+++ cypress/main.ts
@@ -1,5 +1,5 @@
-import { TOKENS, services as e2e } from '../services'
 import { TOKENS as DEV_TOKENS, services as development } from '@/services/development'
+import { TOKENS, services as e2e } from '@/services/e2e'
 import { build } from '@/services/utils'
 
 (async () => {
